gmail_read/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
bs64==0.0.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-environ==0.11.2
djangorestframework==3.15.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
lxml==5.2.1
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycryptodome==3.20.0
pyparsing==3.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.5
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
